requirements-hw.txt
paho-mqtt==2.1.0
smbus2==0.4.3
spidev==3.6
